kbouncer_sim: overhaul printing and INS detection
Printing was updated help users interpret the messages; before
it was a bit esoteric without establishing context. Another
major change is the refinement of the indirect control transfer
instruction detection. Now we have more accuracy in identifying
instructions that will be the end of most gadgets when doing
an instruction walk.